Move utils/ to common/utils/
