+++ apps/client/src/common/utils/auth.ts
