+++ apps/client/src/features/onboarding/api/options.ts
@@ -1,7 +1,7 @@
 import { mutationOptions } from "@tanstack/react-query"
 
 import { client } from "@/common/api/client"
-import { delayWithPromise } from "@/utils/delay"
+import { delayWithPromise } from "@/common/utils/delay"
 
 export const onboardingMutationOptions = (minimumDelayMs: number) =>
     mutationOptions({
