+++ apps/client/src/routes/(onboarding)/onboarding.tsx
@@ -2,10 +2,10 @@ import { useMutation } from "@tanstack/react-query"
 import { createFileRoute, redirect } from "@tanstack/react-router"
 
 import { useEffectOnce } from "@/common/hooks/use-effect-once"
+import { isAuthenticated, isOnboarded } from "@/common/utils/auth"
 import { onboardingMutationOptions } from "@/features/onboarding/api/options"
 import { OnboardingAnimation } from "@/features/onboarding/ui/multi-step-loader"
 import { Page } from "@/libs/ui/page"
-import { isAuthenticated, isOnboarded } from "@/utils/auth"
 
 // (onboarding) authentication & onboarded check
 export const Route = createFileRoute("/(onboarding)/onboarding")({
