+++ apps/client/src/routes/(public)/_layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet, createFileRoute, redirect } from "@tanstack/react-router"
 
-import { isAuthenticated } from "@/utils/auth"
+import { isAuthenticated } from "@/common/utils/auth"
 
 // (public) authentication check
 export const Route = createFileRoute("/(public)")({
